+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 ## Unreleased
+### Changed
+- Modified (some adjustments): volume - titles and descriptions -, metadata volume's - titles and content - are now
+  stored in the T_VOLUMES and T_VOLUME_METADATAS tables respectively.
+
 ### Fixed
 - [RT#187856]: repeated reminders fail.
 ### Added